[Ref] Silence 64-bit warning.
git-svn-id: https://source.openmpt.org/svn/openmpt/trunk/OpenMPT@8514 56274372-70c3-4bfc-bfc3-4c3a0b034d27

+++ soundlib/plugins/PluginMixBuffer.h
@@ -33,7 +33,7 @@ protected:
 	static const uintptr_t additionalBuffer = bufferAlignmentInBytes / sizeof(buffer_t);
 
 	// Return pointer to an aligned buffer
-	buffer_t *GetBuffer(uint32 index) const
+	buffer_t *GetBuffer(size_t index) const
 	//-------------------------------------
 	{
 		MPT_ASSERT(index < inputs.size() + outputs.size());
